chore(connlib): move filters from gateway module to shared module
