+++ rust/connlib/tunnel/src/peer.rs
@@ -4,9 +4,7 @@ use std::time::{Duration, Instant};
 
 use chrono::{DateTime, Utc};
 use connlib_shared::messages::gateway::{ResolvedResourceDescriptionDns, ResourceDescription};
-use connlib_shared::messages::{
-    gateway::Filter, gateway::Filters, ClientId, GatewayId, ResourceId,
-};
+use connlib_shared::messages::{ClientId, Filter, Filters, GatewayId, ResourceId};
 use connlib_shared::DomainName;
 use ip_network::IpNetwork;
 use ip_network_table::IpNetworkTable;
@@ -654,8 +652,7 @@ mod tests {
 
     use chrono::Utc;
     use connlib_shared::messages::{
-        gateway::{Filter, PortRange},
-        ClientId, ResourceId,
+        ClientId, ResourceId, {Filter, PortRange},
     };
     use ip_network::Ipv4Network;
 
@@ -1112,7 +1109,7 @@ mod proptests {
     };
 
     use super::*;
-    use connlib_shared::{messages::gateway::PortRange, proptest::*};
+    use connlib_shared::{messages::PortRange, proptest::*};
     use ip_network::{Ipv4Network, Ipv6Network};
     use ip_packet::make::{icmp_request_packet, tcp_packet, udp_packet};
     use itertools::Itertools;
